refactor: prettify code
