+++ src/operations/docs/types/OwnerType.ts
@@ -1,4 +1,4 @@
 export const OwnerType = {
   SCHOOL: "SCHOOL",
   STUDENT: "STUDENT",
-};
+} as const;
