+++ src/providers/constants.js
@@ -1,0 +1,3 @@
+export const MULTIPART_HEADERS = {
+  "Content-Type": "multipart/form-data",
+};
