+++ src/providers/utils.js
@@ -19,7 +19,3 @@ export const toCamelCaseJSON = (input) => {
     };
   }, {});
 };
-
-export const MULTIPART_HEADERS = {
-  "Content-Type": "multipart/form-data",
-};
